rds_schema_lambda/requirements_test.txt
boto3
psycopg2-binary
pytest
requests
